add specs
