+++ test/Blockable.test.js
@@ -1,5 +1,7 @@
 const Blockable = artifacts.require("Blockable");
 
+const { advanceBlock } = require("./shared/advance_block");
+
 contract("Blockable", (accounts) => {
   let instance;
   let blocksPrecision;
@@ -8,28 +10,6 @@ contract("Blockable", (accounts) => {
   const params = {
     blocksPerEra: 10,
     eraMax: 12,
-  };
-
-  advanceBlock = async (blocksNumber) => {
-    for (let i = 0; i < blocksNumber; i++) {
-      let promise = new Promise((resolve, reject) => {
-        web3.currentProvider.send(
-          {
-            jsonrpc: "2.0",
-            method: "evm_mine",
-            id: new Date().getTime(),
-          },
-          (err, result) => {
-            if (err) {
-              return reject(err);
-            }
-            const newBlockHash = web3.eth.getBlock("latest").hash;
-
-            return resolve(newBlockHash);
-          }
-        );
-      });
-    }
   };
 
   beforeEach(async () => {
