+++ test/ProducerContract.test.js
@@ -4,6 +4,8 @@ const ProducerPool = artifacts.require("ProducerPool");
 const RcToken = artifacts.require("RcToken");
 
 const expectRevert = require("@openzeppelin/test-helpers").expectRevert;
+const { rcTokenDeployed } = require("./shared/rc_token_deployed");
+const { advanceBlock } = require("./shared/advance_block");
 
 contract("ProducerContract", (accounts) => {
   let instance;
@@ -16,28 +18,6 @@ contract("ProducerContract", (accounts) => {
     await instance.addProducer(10, name, "photoURL", "135465-005", { from: address });
   };
 
-  advanceBlock = async (blocksNumber) => {
-    for (let i = 0; i < blocksNumber; i++) {
-      let promise = new Promise((resolve, reject) => {
-        web3.currentProvider.send(
-          {
-            jsonrpc: "2.0",
-            method: "evm_mine",
-            id: new Date().getTime(),
-          },
-          (err, result) => {
-            if (err) {
-              return reject(err);
-            }
-            const newBlockHash = web3.eth.getBlock("latest").hash;
-
-            return resolve(newBlockHash);
-          }
-        );
-      });
-    }
-  };
-
   const producerPoolArgs = {
     totalTokens: "750000000000000000000000000",
     halving: 50,
@@ -46,7 +26,7 @@ contract("ProducerContract", (accounts) => {
   };
 
   beforeEach(async () => {
-    rcToken = await RcToken.new("1500000000000000000000000000");
+    rcToken = await rcTokenDeployed();
 
     userContract = await UserContract.new();
