+++ test/ResearcherPool.test.js
@@ -2,43 +2,22 @@ const RcToken = artifacts.require("RcToken");
 const ResearcherPool = artifacts.require("ResearcherPool");
 
 const expectRevert = require("@openzeppelin/test-helpers").expectRevert;
+const { rcTokenDeployed } = require("./shared/rc_token_deployed");
+const { advanceBlock } = require("./shared/advance_block");
 
 contract("ResearcherPool", (accounts) => {
   let instance;
   let [owner, researcher1Address, researcher2Address] = accounts;
 
   const args = {
-    totalRcTokens: "1500000000000000000000000000",
     totalResearcherPoolTokens: "30000000000000000000000000",
     halving: 12,
     totalEras: 96,
     blocksPerEra: 12,
   };
 
-  advanceBlock = async (blocksNumber) => {
-    for (let i = 0; i < blocksNumber; i++) {
-      let promise = new Promise((resolve, reject) => {
-        web3.currentProvider.send(
-          {
-            jsonrpc: "2.0",
-            method: "evm_mine",
-            id: new Date().getTime(),
-          },
-          (err, result) => {
-            if (err) {
-              return reject(err);
-            }
-            const newBlockHash = web3.eth.getBlock("latest").hash;
-
-            return resolve(newBlockHash);
-          }
-        );
-      });
-    }
-  };
-
   beforeEach(async () => {
-    const rcToken = await RcToken.new(args.totalRcTokens);
+    const rcToken = await rcTokenDeployed();
     instance = await ResearcherPool.new(rcToken.address, args.halving, args.totalEras, args.blocksPerEra);
 
     await instance.newAllowedCaller(owner);
